refactor: Move <ConversationsHistoryList/> data logic to <ConversationsHistorySidebar/>
